Optimize: Modularize the code
